Domain file parsing complete. Working on problem file parsing.
